+++ src/ThreadElements/Shots.java
@@ -125,7 +125,7 @@ public class Shots extends Thread{
 
             // Criando margens do icone 
             this._IconBounds = new Rectangle();
-            this._IconBounds.setBounds(this._OriginPoint.getX(), this._OriginPoint.getY(), 
+            this._IconBounds.setBounds((int)(this._OriginPoint.getX()), (int)(this._OriginPoint.getY()), 
                                        this._UpdatedLocation.getSizeX(), this._UpdatedLocation.getSizeY());
 
             // Definindo tamanho 
@@ -163,8 +163,8 @@ public class Shots extends Thread{
         for (SpritePoint spritePoint : _TrajectoryPoints) {
             if (!this._Hit) {
                 this._UpdatedLocation.setPositions(spritePoint);
-                this._IconBounds.x = this._UpdatedLocation.getX();
-                this._IconBounds.y = this._UpdatedLocation.getY();
+                this._IconBounds.x = (int)(this._UpdatedLocation.getX());
+                this._IconBounds.y = (int)(this._UpdatedLocation.getY());
                 try {
                     Thread.sleep(this._UpdteFrequency);
                 } catch (InterruptedException e) {
